team: don't log failures to connect to teamd on error level
ensure_teamd_connection() is called from multiple spots. Sometimes
we call opportunistically without having started teamd (e.g. when on
update_connection() when generating a connection for teaming device that
was created) and handle the failure to connect gracefully.

Let's not pollute the logs with things on ERROR level that are not
actually serious. Demote the logging statement to DEBUG and log an ERROR
only when we expect ensure_teamd_connection() to actually succeed.