glib-aux: add NMPtrArray helper
Add a simple structure that can own a list of pointers.
The main purpose is to return one or more pointers from a function,
where ownership gets transferred.

This is a more limited form of a GPtrArray. Except, it actually NULL
terminates the list, and it requires only one heap allocation for
the structure and the list.

With using NMPtrArrayStack, it even can track the first 8 pointers (in x64_86)
on the stack and avoiding heap allocation altogether. That is of course
only useful, when the ptr array is not to be returned from the function
(to the caller), but instead passed down to called functions which can
fill it.