platform: fix nmp_object_copy(id_only) for object that don't implement cmd_plobj_id_copy()
The if-else-if was wrong. It meant that if an object did not implement
cmd_plobj_id_copy(), nothign was copied (for id-only).

I think this code path was not actually hit, because we never clone
an object only by ID.

Fixes: c91a4617a102 ('nmp-object: allow missing implementations for certain virtual functions')
(cherry picked from commit ee34eeafb9c4379828e3ac6c5b279ca9426bd99e)