glib-aux: drop unused _nm_dbus_connection_call_finish()
This wasn't used. It also doesn't make much sense, because
g_dbus_connection_call() accepts a reply type argument. So if you want
to check the reply type, you can do so when making the call.

That differs from g_dbus_proxy_call(), where this might make some sense
and is actually used.

While at it, also drop _nm_dbus_typecheck_response() from the header
file. It seems not general purpose enough and is not necessarily a great
pattern. Because,

- for g_dbus_connection_call() you either specify the reply_type when making
  the call, or you do some more elaborate handling of the result -- not merely
  _nm_dbus_typecheck_response().

- for g_dbus_proxy_call(), you can use _nm_dbus_proxy_call_finish() to
  get it, or again, you want some non-trivial type checking.