core: set routing rule protocol properly
When creating NMPlatformRoutingRule from NMIPRouteRule object, the
protocol is not being set properly. According to linux kernel
documentation FRA_PROTOCOL indicates the originator of the rule.

In this case the route rule is coming from a connection and therefore
the originator of the rule is the user. The correct value is
RTPROT_STATIC which means the rule is installed by the administrator.

Fixes: 3f9347745b30 ('core: add handling of IP routing rules to NMDevice')

Signed-off-by: Fernando Fernandez Mancera <ffmancera@riseup.net>