platform: rework parsing of objects in _rtnl_handle_msg()
Most rtnetlink messages directly correspond to one NMPObject.
Not so IPv6 routes (RTM_{NEW,DEL}ROUTE), which can have several
next-hops, and for each next-hop we parse a separate NMPObject.

Previously, _rtnl_handle_msg() did that parsing of multiple messages
in its own loop. Move that code to a separate function
nmp_object_new_from_nl_v(). This function now fills a NMPtrArray
and can return the entire list at once.

NMPtrArray is used because it's a lightweight structure that can own
a list of pointers (exactly as needed here). Also, in the vast majority
of cases, there will be only one object returned, so the caller
(_rtnl_handle_msg()) can start with a stack-allocated NMPtrArrayStack,
which holds space for some pointers before requiring the first
allocation. Thanks to cleanup attributes that own the references, this is
relatively clear code and should have basically no additional overhead.

The reason for doing this is that with `ip route replace`, kernel will replace
the entire IPv6 multi-hop route (which in NMPlatform corresponds to
multiple single-hop objects). nmp_cache_update_netlink_route() is aware of
that and tries to find which routes were replaced. But to do that
correctly, it will need the complete list of all the new objects. Now we
first parse that list so we have it upfront. This change will be
important later.