platform: rework parsing of objects in _rtnl_handle_msg()
Most rtnetlink messages directly correspond to one NMPObject.
Not so IPv6 routes (RTM_{NEW,DEL}ROUTE), which can have several
next-hops, and for each next-hop we parse a separate NMPObject.

Previously, _rtnl_handle_msg() did that parsing of multiple messages
in its own loop. Move that code to a separate function
nmp_object_new_from_nl_v(). This function now fills a NMPtrArray
and can return the entire list at once.

This seems nicer code, instead of embedding the loop inside
_rtnl_handle_msg() which already does a lot of non-trivial things.

It also has the benefit, that we have the entire data parsed at once.
Maybe this could be useful to nmp_cache_update_netlink_route(), to know
what routes will be added.

One potential downside is that we now parse everything earlier, although
the code below could determine after the first object that a resync is
required, and the remaining objects are not relevant anymore. But that
does not justify making embedding the parsing in _rtnl_handle_msg().

NMPtrArray is used because it's a lightweight structure that can own
a list of pointers (exactly as needed here). Also, in the vast majority
of cases, there will be only one object returned, so the caller
(_rtnl_handle_msg()) can start with a stack-allocated NMPtrArrayStack,
which holds space for some pointers before requiring the first
allocation. Thanks to cleanup attributes that own the references, this is
relatively clear code and should have basically no additional overhead.