device: remember whether the DHCP client-id/duid is a fixed value
Remember when the client-id is set to a explicit fixed value, so that
we can take decision based on that (such as, whether the DHCP client
can be restarted without the risk that the lease changes).

This is not perfect because for client-id we accept also "stable",
which may also translate to a fixed value depending on the value of
another property. For now, only evaluate whether the properties
contain explicit fixed values.