core/team: do not care about who owns the bus name
We don't need to care about the processes, just the bus name.

We also don't need to kill out process, it's just going to die by itself
if it fails to grab the bus name. On top of that, the process that grabbed
the name might not even be the same as we spawned, there could have been
a fork in between. Don't assume too much.